refactor: Extract dashboard HTML to separate module
Split large inline HTML into lib/dashboard/trinity-html.ts module for:
- Better code organization
- Smaller route handler file
- Easier maintenance and testing
- Improved linting compliance
